pyproject.toml
[project]
name = "docker-image-pusher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "prettytable>=3.16.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.0",
    "typer>=0.15.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "playwright>=1.40.0",
]

[project.scripts]
cli = "src.main:app"

[tool.uv]
package = true
